add ncurses animations

+++ src/StatisticsGenerator.cpp
@@ -10,7 +10,7 @@ void StatisticsGenerator::run()
 {
     while(app_running)
     {
-        correct_percent = static_cast<double>(correct_decoding / number_of_operations);
+        correct_percent = static_cast<double>(correct_decoding) / static_cast<double>(number_of_operations);
         std::this_thread::sleep_for(std::chrono::seconds(1));
     }
 }
